refactor(bank-leumi): split bank leumi into stateless units
