+++ src/constants.js
@@ -25,6 +25,7 @@ export const LOGIN_RESULT = {
   INVALID_PASSWORD: 'INVALID_PASSWORD',
   CHANGE_PASSWORD: 'CHANGE_PASSWORD',
   UNKNOWN_ERROR: 'UNKNOWN_ERROR',
+  INVALID_OPTIONS: 'INVALID_OPTIONS',
 };
 
 export const NAVIGATION_ERRORS = {
