+++ src/helpers/errors.js
@@ -1,0 +1,10 @@
+import { GENERAL_ERROR } from '../constants';
+
+function createGeneralError() {
+  return {
+    success: false,
+    errorType: GENERAL_ERROR,
+  };
+}
+
+export default createGeneralError;
