+++ src/scrapers/leumi/definitions.js
@@ -1,0 +1,3 @@
+const BASE_URL = 'https://hb2.bankleumi.co.il';
+
+export default BASE_URL;
