+++ src/scrapers/leumi/scrape-transactions.js
@@ -1,0 +1,211 @@
+import moment from 'moment';
+import {
+  dropdownSelect,
+  dropdownElements,
+  fillInput,
+  clickButton,
+  waitUntilElementFound,
+  pageEvalAll,
+  elementPresentOnPage,
+} from '../../helpers/elements-interactions';
+import { navigateTo, waitForNavigation } from '../../helpers/navigation';
+import { SHEKEL_CURRENCY, NORMAL_TXN_TYPE, TRANSACTION_STATUS } from '../../constants';
+import BASE_URL from './definitions';
+import createGeneralError from '../../helpers/errors';
+
+const DATE_FORMAT = 'DD/MM/YY';
+
+function getTransactionsUrl() {
+  return `${BASE_URL}/ebanking/Accounts/ExtendedActivity.aspx?WidgetPar=1#/`;
+}
+
+function getAmountData(amountStr) {
+  const amountStrCopy = amountStr.replace(',', '');
+  const amount = parseFloat(amountStrCopy);
+  const currency = SHEKEL_CURRENCY;
+
+  return {
+    amount,
+    currency,
+  };
+}
+
+function convertTransactions(txns) {
+  return txns.map((txn) => {
+    const txnDate = moment(txn.date, DATE_FORMAT).toISOString();
+
+    const credit = getAmountData(txn.credit).amount;
+    const debit = getAmountData(txn.debit).amount;
+    const amount = (Number.isNaN(credit) ? 0 : credit) - (Number.isNaN(debit) ? 0 : debit);
+    return {
+      type: NORMAL_TXN_TYPE,
+      identifier: txn.reference ? parseInt(txn.reference, 10) : null,
+      date: txnDate,
+      processedDate: txnDate,
+      originalAmount: amount,
+      originalCurrency: SHEKEL_CURRENCY,
+      chargedAmount: amount,
+      status: txn.status,
+      description: txn.description,
+      memo: txn.memo,
+    };
+  });
+}
+
+async function extractCompletedTransactionsFromPage(page) {
+  const txns = [];
+  const tdsValues = await pageEvalAll(page, '#WorkSpaceBox #ctlActivityTable tr td', [], (tds) => {
+    return tds.map(td => ({
+      classList: td.getAttribute('class'),
+      innerText: td.innerText,
+    }));
+  });
+
+  for (const element of tdsValues) {
+    if (element.classList.includes('ExtendedActivityColumnDate')) {
+      const newTransaction = { status: TRANSACTION_STATUS.COMPLETED };
+      newTransaction.date = (element.innerText || '').trim();
+      txns.push(newTransaction);
+    } else if (element.classList.includes('ActivityTableColumn1LTR') || element.classList.includes('ActivityTableColumn1')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.description = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('ReferenceNumberUniqeClass')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.reference = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('AmountDebitUniqeClass')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.debit = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('AmountCreditUniqeClass')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.credit = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('number_column')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.balance = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('tdDepositRowAdded')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.memo = (element.innerText || '').trim();
+      txns.push(changedTransaction);
+    }
+  }
+
+  return txns;
+}
+
+async function extractPendingTransactionsFromPage(page) {
+  const txns = [];
+  const tdsValues = await pageEvalAll(page, '#WorkSpaceBox #trTodayActivityNapaTableUpper tr td', [], (tds) => {
+    return tds.map(td => ({
+      classList: td.getAttribute('class'),
+      innerText: td.innerText,
+    }));
+  });
+
+  for (const element of tdsValues) {
+    if (element.classList.includes('Colume1Width')) {
+      const newTransaction = { status: TRANSACTION_STATUS.PENDING };
+      newTransaction.date = (element.innerText || '').trim();
+      txns.push(newTransaction);
+    } else if (element.classList.includes('Colume2Width')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.description = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('Colume3Width')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.reference = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('Colume4Width')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.debit = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('Colume5Width')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.credit = element.innerText;
+      txns.push(changedTransaction);
+    } else if (element.classList.includes('Colume6Width')) {
+      const changedTransaction = txns.pop();
+      changedTransaction.balance = element.innerText;
+      txns.push(changedTransaction);
+    }
+  }
+
+  return txns;
+}
+
+async function fetchTransactionsForAccount(page, startDate, accountId) {
+  await dropdownSelect(page, 'select#ddlAccounts_m_ddl', accountId);
+  await dropdownSelect(page, 'select#ddlTransactionPeriod', '004');
+  await waitUntilElementFound(page, 'select#ddlTransactionPeriod');
+  await fillInput(
+    page,
+    'input#dtFromDate_textBox',
+    startDate.format(DATE_FORMAT),
+  );
+  await clickButton(page, 'input#btnDisplayDates');
+  await waitForNavigation(page);
+  await waitUntilElementFound(page, 'table#WorkSpaceBox table#ctlActivityTable');
+
+  const hasExpandAllButton = await elementPresentOnPage(page, 'a#lnkCtlExpandAllInPage');
+
+  if (hasExpandAllButton) {
+    await clickButton(page, 'a#lnkCtlExpandAllInPage');
+  }
+
+  const selectedSnifAccount = await page.$eval('#ddlAccounts_m_ddl option[selected="selected"]', (option) => {
+    return option.innerText;
+  });
+
+  const accountNumber = selectedSnifAccount.replace('/', '_');
+
+  const pendingTxns = await extractPendingTransactionsFromPage(page);
+  const completedTxns = await extractCompletedTransactionsFromPage(page);
+  const txns = [
+    ...pendingTxns,
+    ...completedTxns,
+  ];
+
+  return {
+    accountNumber,
+    txns: convertTransactions(txns),
+  };
+}
+
+async function fetchTransactions(page, startDate) {
+  const res = [];
+  // Loop through all available accounts and collect transactions from all
+  const accounts = await dropdownElements(page, 'select#ddlAccounts_m_ddl');
+  for (const account of accounts) {
+    // Skip "All accounts" option
+    if (account.value !== '-1') {
+      res.push(await fetchTransactionsForAccount(page, startDate, account.value));
+    }
+  }
+  return res;
+}
+
+
+async function scrapeTransactions(page, options) {
+  try {
+    const defaultStartMoment = moment().subtract(1, 'years').add(1, 'day');
+    const startDate = options.startDate || defaultStartMoment.toDate();
+    const startMoment = moment.max(defaultStartMoment, moment(startDate));
+
+    const url = getTransactionsUrl();
+    await navigateTo(page, url);
+
+    const accounts = await fetchTransactions(page, startMoment);
+
+    return {
+      success: true,
+      accounts,
+    };
+  } catch (error) {
+    return createGeneralError();
+  }
+}
+
+export default scrapeTransactions;
